+++ jfreechart/src/test/java/chart/pie/JFreePieChartTest.java
@@ -20,7 +20,7 @@ public class JFreePieChartTest extends AbstractChartTest<PieChart> {
         String text = "How Much Wood Would A Woodchuck Chuck If A Woodchuck Could Chuck Wood ?";
         Map<String, Long> words = Stream.of(text.split(" ")).collect(Collectors.groupingBy(k -> k, Collectors.counting()));
 
-        words.forEach((key, value) -> chart.putEntries(key, NumberDatasetFactory.create(value)));
+        words.forEach((key, value) -> chart.putEntries(key, NumberDatasetFactory.create(value.doubleValue())));
 
         save("PieChart");
     }
@@ -31,7 +31,7 @@ public class JFreePieChartTest extends AbstractChartTest<PieChart> {
         String text = "How Much Wood Would A Woodchuck Chuck If A Woodchuck Could Chuck Wood ?";
         Map<String, Long> words = Stream.of(text.split(" ")).collect(Collectors.groupingBy(k -> k, Collectors.counting()));
 
-        words.forEach((key, value) -> chart.putEntries(key, NumberDatasetFactory.create(value, Optional.of(new Color(key.hashCode())))));
+        words.forEach((key, value) -> chart.putEntries(key, NumberDatasetFactory.create(value.doubleValue(), Optional.of(new Color(key.hashCode())))));
 
         save("PieChartColor");
     }
